Fixing some bugs and adding features
tmframework.lua: Fixed some bugs and added UUID and key generation to the begin of the library. so when loaded it will automatically gather its dependencies and generate an uuid and sha-256 hash, this is useful for on computer backups. I will probably add some feature to the client to load config files.